feat(PriceClient): Multiple price feeds
This feature adds a new class named PriceClient, which acts like an
aggregator for a user-defined list of price feeds. This permits prices
to be sourced from a range of sources according to preference (i.e.
failover to a lesser-preferred source in the event that a more-preferred
source is unavailable). This should improve the robustness of price
retrieval, helping to relieve CoinGecko Pro request load, as well as
CoinGecko free-tier rate limiting in the Across v2 bots.

This feature benefits from some of the functionality that was recently
added to the Coingecko class:
 - Basic support for caching of prices.
 - Basic support for bundling token price requests.

Ultimately, that functionality should be removed from the existing
Coingecko class once all users have migrated across to use the
PriceClient.