improve(PriceClient): Supply price feeds during init
Per feedback from @daywiss, the PriceClient implementation can be
simplified a little bit by supplying the list of price feeds during
instantiation.

This however makes the singleton get() method more complicated, and the
caller of a get() method would have to separately verify that their
supplied price feed was actually used (or ignored in the event that the
PriceClient had previously been initialised). Therefore, remove the
singleton pattern entirely. This does permit multiple instances of
PriceClient to be instantiated, but that doesn't appear to be a common
pattern in our code and it can be managed by the caller(s).