feat(gasPriceOracle): New helper for estimating gas prices
getGasPriceEstimate() provides a gas price estimate for a range of
chains. A chain-specific function is called to perform the right lookup,
and estimates are normalised to Type 2 (EIP-1559) formatting. If a chain
is not recognised, the caller can decide whether to fall back to a
legacy/type 0 lookup (i.e. eth_gasPrice).

Callers can extract the total price GasPriceEstimate.maxFeePerGas. If a
priority fee is in effect on the network, then it can be obtained from
the field GasPriceEstimate.maxPriorityFeePerGas.