improve(utils): Improve gas price estimation (#102)
Testing has shown that eth_getPrice() consistently returns a value that
excludes the chain priority fee. This means that gas price estimation is
currently 1.5 Gwei lower than it should be (ethers defaults to assuming
1.5 Gwei for maxPriorityFeePerGas).

This change introduces getGasPriceEstimate(), to provide a gas price
estimate for a range of chains. A chain-specific function is called to
perform the right lookup, and estimates are normalised to Type 2 (EIP-
1559) formatting. If a chain is not recognised, the caller can decide
whether to fall back to a legacy/type 0 lookup (i.e. eth_gasPrice).