feat!: Add config options to discount fees by destination chain ID
Adds a required param to relayerFeeCalculator's `relayerFeeDetails` function: `destinationChainId`. (I can see an argument for this being optional but the codes a bit cleaner to require it as optional and the way its used today, it should be easy to upgrade to the next SDK version and change the `relayerFeeDetails` call.

This PR redefines the interface for `gasDiscountPercent` and `capitalDiscountPercent` in the `RelayFeeCalculatorConfig` rather than add new properties. The reason I do this is because neither the Relayer or the Frontend uses these properties currently, and updating these interfaces in-place required less code than introducing new ons.