chore(UBA): Tweak outflowIsFill predicate
Incorporate some fixes from Nick's across-2.5 branch, whilst also
adjusting the outflowIsFill predicate to avoid the use of isSlowRelay.

isSlowRelay was an unfortunate choice, because it gets relocated as
part of the new SpokePool migration. Instead, use repaymentChainId
because it's unique to a fill and is not impacted by the new ABI.