feat(UBA): :zap: optimize uba fee calculation's spoke pool clients
This change adds the concept of a UBAFeeSpokeStore which stores information needed for one half of the computation values of a UBAFee Calculation. Two of these stores can be combined using the UBAFeeCalculator wrapper to facilitate the calculation of a valid UBA Fee.

Breaking Change: The class definition of the UBAFeeCalculator has changed.