feat(UBA): :sparkles: add hurdle balance threshold trigger in historical balance calculations
This change makes the running balance aware of the concept of the trigger balance threshold. To cap the organic growth of a given spokepool's running balance, a mechanism for preventing increasing balances has been proposed. As the running balance is calculated, if the balance exceeds a certain threshold dictated by the on-chain AcrossConfigStore then this balance will be artificially lowered by the ACX dataworker. As such, this non-event based event must be captured in the running balance to maintain a uniform understanding of the running balances across all ecosystem relayers/bots/UIs