feat(UBA): Support overloaded ExecutedRootBundle runningBalances
The UBA needs to keep track of the size of the incentive pool for each
token on each chain. Per the pending updates to UMIP-157, this will be
done by appending the running incentive balances to the
ExecutedRootBundle runningBalances array.

This change permits the HubPoolClient to infer the presence of incentive
balances in the runningBalances array based on its length.