improve: Split ConfigStoreClient update() (#227)
This change provides an option to override ConfigStoreClient on-chain queries
via _update(). The pattern is pretty consistent with what we have in the
HubPoolClient and SpokePoolClient. For the MockConfigStoreClient it was
necessary to inherit super._update() by default, for compatibility with existing
uses of the MockConfigStoreClient in relayer-v2 tests. Ultimately it would be
nice to remove that.

Addtionally, support mocked getBlock() for HubPoolClient test. This is needed
for the pending runningBalances UBA update.