fix: Remove off-by-one error in SpokePoolClient (#231)
The SpokePool "numberOfDeposits" state variable is the depositId of the next
deposit to be confirmed; not the last one.

I'm not aware of any cases where this has caused an error, but the bots sometimes
log that they can't find a specific deposit during periods of RPC instability. Fixing
this allows us to rule it out as a potential error source.