fix(mock): Don't override generated fill destinationToken
The SpokePoolClient tries to resolve the destinationToken against the
set of known tokens in the HubPoolClient, so it's important to propagate
the value supplied by the caller. If no value has been supplied, sub in
the ZERO_ADDRESS, because this fits with the scenario where the
SpokePoolClient has no reference for the HubPoolClient.