feat(gasPriceOracle): Support Polygon Gas Station (#241)
Prompted by the migration from matic.network to polygon.technology
today, add support for the Gas Station API and add a fallback strategy
in case the gas station fails to resolve.

This PR expanded into a minor re-org of the gasPriceOracle, and also
results in enabling EIP-1559 for Arbitrum. In practice it probably makes
little difference to do this, since Arbitrum accepts EIP-1559
transaction types but refunds the priority fee (per the Nitro upgrade).