fix: Revert to legacy gas pricing for Arbitrum (#245)
Ethers hardcodes a price of 1.5 Gwei for the priority fee. On Arbitrum, the
priority fee is refunded to the caller, so it doesn't makse sense to use that
in gas computations since it results in over-estimation.

Additionally, I spotted a couple of minor issues during CI, so fix those as well.