fix(gasPriceOracle): Assorted fixes (#246)
I found several issues in the gasPriceOracle:

Polygon gas station lookups are failing because the request resolves to
//v2 instead of /v2. Polygon don't automatically squash the additional /
characters, and instead respond with a 404. This hasn't impacted
service because the adapter falls back to an on- chain price lookup,
but it delays gas price resolution and probably gives less accurate
results.

After resolving the gas station issue, the oracle.e2e test needed fixing
because the Polygon adapter no longer returns mocked gas prices.

As part of the updates to the oracle.e2e test, I realised that the use 
of assert.ok() is node-specific and seems irrelevant to jest. So, use 
expect() instead. This makes the tests a bit more compact and easier to
read.

Additionally:

An Arbitrum-specific adapter has been added that effectively nulifieies
the priority fee. It's set to 1 Wei because a priority fee of 0 Wei is
used to signal a legacy (type 0) transaction.

As part of this change I have also updated the RPC list used when local
RPCs are not defined. These are the current high-performing publicly
availably RPCs, as reported by chainlist.org.