fix(UBA): Fix RefundRequest validation
This was wrong from the beginning because it didn't properly validate
the Deposits. It also needs to be async because it must be able to
search for old fills and deposits via an RPC provider.

While here, remove refundRequestIsValid() from UBAClientBase because
it's unlikely to be used in its current form, and refund request
validation belongs between the UBAClient and SpokePoolClients.