fix(UBA): Fix RefundRequest validation (#272)
Refund Request validation has been wrong since it was first added,
because it relied on the deposits having previously been validated, and 
that wasn't the case. It also needs to be async because it must be able
to search for old fills and deposits via an RPC provider.

While here, remove refundRequestIsValid() from UBAClientBase because
it's unlikely to be used in its current form, and refund request
validation belongs between the UBAClient and SpokePoolClients.