improve(test): Singleton event manager in mocked clients
This change migrates to the use of a singleton for event managers,
enabling multiple "client" classes to see consistent block number
increments. This addresses a race condition in the relayer-v2
runningBalances test where the ConfigStoreClient and HubPoolClient can
get out of sync and disagree about the applicable version.

While here, add an additional check to ensure that FundsDeposited events
are generated with sequential (or at least incrementing) depositIds.