fix: Subtract netRunningBalanceAdjustment when over target
If we are over the target, subtract the difference from the net running balance adjustment so that the dataworker can instruct the spoke pool to return funds to the hub pool.

If we are under the target, add the difference to the net running balance adjustment so that the dataworker can instruct the hub pool to send funds to the spokepool.