fix: Restore SpokePoolClient _update pattern
The intention of the _update() method is that it supplies all
network-related responses. Querying the RPC provider from update()
breaks this pattern and will effectively break the MockSpokePoolClient
with mocked events.

So, query all block data from _update() instead. This has the side
advantage of allowing all blocks to be queried in parallel, and we can
collate and filter duplicates in advance.