fix: Pass in chainIdIndices instead of enabledChainIds
There is a very important difference between calling `getEnabledChains` and `getChainIdIndices` on the `ConfigStoreClient`. The latter is used to index block numbers to chains while the former is used much more seldomly to limit which chains we care about for whatever operation.

This PR fixes critical bugs in the various clients that create block ranges. I discovered this bug while integration testing the dataworker and UBA client