improve(gasPriceOracle): Initial zkSync support
This change explicitly signals that zkSync gas price estimates should be
based on the legacy method, rather than eip1559.

The zkSync docs indicate that type 2 (EIP1559) transactions are
supported, but clarify that the maxFee* parameters are ignored:
  https://era.zksync.io/docs/reference/concepts/transactions.html#eip-1559-0x2

The zkSync-specific provider references the ethers getGasPrice() method,
rather than the more post-1559 getFeeData() method that supports the
priority fee:
  https://era.zksync.io/docs/api/js/providers.html#getgasprice

L2s like zkSync and Optimism are additionall indicating more dynamic
pricing based on the corresponding L1 cost of the transaction, and this
requires fee estimation based on the actual transaction request.
Ultimately we may need to support to a strategy like this in order to
get accurate gas price estimates.