fix(UBA): :bug: assert that config store is updated
In order to find the list of enabled chains from the config store, we need to ensure that the config store has been updated. This is something that should be provided prior to the config store being passed into the constructor of the UBAClient.

Breaking Changes: Any code with the assumption that an unupdated config store can be passed into the UBAClientWithRefresh