fix: Better gas estimation for Base
It was reported a few weeks ago that a gas spike on mainnet caused
deposits destined for Base to be underpriced; I couldn't figure out why
this would be the case for Base but not Optimism, but this may have
contributed.

As part of this, add a simple utility to indicate whether a chain ID is
an OP stack implementation.