improve(HubPoolClient): Process LP fees in batch
This change removes a source of significant inefficiency in the
HubPoolClient. The pre-existing implementation processes each Deposit's
realizedLpFeePct independently, but in parallel. In cases where there
are n deposits with the same quoteTimestamp, this results in n
concurrent searches for the underlying block number. Additionally, when
there are n deposits with matching originToken and quoteTimestamp (as
can easily happen with the popular tokens like (W)ETH and USDC) this
results in similarly duplicate queries for the "opening" HubPool
utilization at the resolved block number.

This commit updates the implementation as follows:
 - Pass the set of Deposits to be computed by the HubPoolClient as an
   array, rather than individually.
 - Filter quoteTimestamps for uniqueness before resolving the
   corresponding HubPool block number.
 - Map HubPool tokens to quoteTimestamps and eventually block numbers in
   batch.
 - Query "opening" HubPool utilization for each HubPool token at the
   relevant block once, and reference the result later for each unique
   deposit.

Debug logging added indicates a significant drop in the number of
HubPool block numbers that need to be resolved; this is especially the
case on popular deposit chains like zkSync and Arbitrum.