improve(SpokePoolClient): Speedup algorithm for appending speed up data to deposits
This PR optimizes compute time for the case where there is a speed up deposit in a SpokePoolClient update.

Currently the algorithm for adding speed up data to deposits is:
- load all speed up events into memory
- loop through all deposits already stored in memory, detect if there is a speed up event newly saved into memory for the deposit, and update the deposit

We needlessly loop through all deposits since there are expected to very few speed ups in a given lookback window relative to the number of deposits loaded. Instead, the proposed algorithm is:
- load all speed up events into memory
- for each new speed up data newly saved, detect if there is a deposit data stored in memory matching the speed up, and update the deposit