fix(BlockUtils): Resolve major BlockFinder inefficiency
The existing BlockFinder implementation can perform very poorly on
chains where the block time has trended lower over time. This is
triggered by the averageBlockTime() function computing the average block
time over a relatively recent history. This recent average is then
extrapolated out over the entire history of the chain by the
BlockFinder.

In cases where the recent block production time is fast, but the
historical block production time was relatively slower, this can lead to
a vast over-estimation of the number of blocks to "look back", and in
testing has resulting in the BlockFinder electing 0 as its starting
block.

This commit makes two changes:

- When estimating the average block time for the purpose of determining
  how far to look back, compute the average over the entire history of
  the chain.
- Increase the initial lookback "cushion" from 1 to 2. This will cause
  the BlockFinder to cast back further to rapidly resolve a block that
  lies before the requested timestamp, but ideally not too far away from
  it.

In testing, the above adjustments result in a significant speed-up in
the relayer, and it's now possible to complete a run on testnets within
~35 seconds.