fix(HubPoolClient): Avoid utilization cache collisions
The last change to HubPool utilization computation ended up reusing the
same cache key for post-relay utilization, but storing the value with a
different format (1 number instead of 2 comma-separated).

This PR replaces the existing getUtilization() function because it's no
longer used, and adds a temporarily change to the post-relay utilization
key to avoid collisions until the original keys are ejected.

This was caught during final local testing of the PR to bump sdk-v2
within relayer-v2.