improve(Clients): Set latestBlockNumber equal to latest block searched and remove redundant latestBlockSearched
Downstream users of the HubPoolClient and the SpokePoolClient expect that the `latestBlockNumber` is the latest block seen by the appropriate providers. However, this is not the case if the clients' `searchConfig.toBlock` lags the HEAD block on the chain, which is possible in production especially since SpokePoolClient updates take a long time to run and are done in parallel.