feat: Simplify HubPoolClient interface for returning L1 --> L2 token mappings (#426)
This PR simplifies the interface used for mapping between HubPool and
SpokePool tokens. The existing constraint of applying the latest token
mapping is retained. Both relayer and dataworker will use the latest
applicable destinationToken as their references for fills and repayments
accordingly.

See also the applicable PR for further discussion and commentary:
https://github.com/across-protocol/sdk-v2/pull/426

Signed-off-by: nicholaspai <npai.nyc@gmail.com>
Co-authored-by: Paul <108695806+pxrl@users.noreply.github.com>