feat: Introduce v3 types (#495)
This PR shuffles some existing types around to make room for the new v3
v3 types. It also introduces some convenience functions that will be
useful for discriminating between v2 and v3 types, as well as extracting
data from them. The v3 types are made available by this change, but are
not imposed on any of the underlying implementation. No functional
change intended.

Fixes ACX-1671