refactor: Remove UBA functionality (#515)
This is complicating the v3 upgrade path. Remove it now; it can be
restored post-migration when v2 support has been removed.

nb. This is a huge diff, and in practice all of this functionality is
unused in the current system. It feels somewhat violent to be removing
this now, but it does also represent a risk to have so much deactivated
code in the codebase.