refactor: Re-organise types around RelayData (#528)
This PR makes two changes:

- Reorganise common type definitions around v*RelayData types.
- Rename type definitions v* -> V* to fit type style norms.

The former is done out of recognition that RelayData definitions
actually drive a lot of the activity in the bots, and this is even
moreso the case with v3. Anchoring types around RelayData definitions
shortens the set of definitions needed in the SpokePool interfaces file,
and also gives assurance that the bots can correctly derive a RelayData
object out of a Deposit, Fill, or even a SlowFillRequest.