feat: Unify v2 & v3 types (#501)
This PR rolls out v3 types everywhere by unifying them under the
Deposit, Fill, ... types that are handled throughout the SDK and relayer
codebases. With this change, all consumers of these datatypes must have
implemented specific handling to discriminate between v2 and v3 events.
In many cases these objects are passed transparently though functions
because they are not evaluating any of the fields which are different
between the two. tsc has been helpful to identify which codepaths need 
to be upgraded.

Co-authored-by: Matt Rice <matthewcrice32@gmail.com>