improve(HubPoolClient): Remove originChain LP fee constraint
Remove a constraint in batch LP fee computation by permitting any
unsorted array of deposit objects, rather than requiring the array be
filtered for a single origin chain. This turned out to be a fairly
simple change, and allows an actor (i.e. the relayer) to batch-compute
the LP fee for a large array of deposits.

A follow up change will update the function prototype to accept a
repaymentChainId.