improve(SpokePoolClient): Remove unsafe function
getExecutedRefunds unintentionally might return unexpected data if the passed in `relayerRefundRoot` is a duplicate between root bundles, which is rare but possible.

This function is only called currently in the BundleDataClient when computing pending bundle refunds (and used to mark off refunds that have been executed) and the validateRunningBalances script, so its impact is limited.

I propose reimplementing this function in the BundleDataClient where it is used so it is not unintentionally relied on in other functions in the future.