improve(SpokePoolClient): Support update timeout (#611)
Allow the caller to specify a timeout on SpokePoolClient.update(). The 
caller can then use SpokePoolClient.isUpdated to determine whether the 
update was successful.

This change also catches RPC errors within SpokePoolClient._update() and
instead produces a log message, subsequently setting 
SpokePoolClient.isUpdated to false. This permits the caller to decide 
how to proceed. The downstream relayer is already robust to this, and 
the dataworker has been updated to require that all SpokePoolClient 
instances succeeded in their updates.