improve(HubPoolClient): Don't throw on token resolution failure
This is too violent for a lot of use cases and requires callers to wrap
their calls in try/catch statements, which has its own set of problems.
Instead, just return undefined if the requested token can't be resolved
and allow the caller to decide on how to respond to that.