refactor: Replace ethers Event with Log
The ethers Event type is an extension of the underlying Log type, which
is actually the type that contains almost all of the event data that the
SDK needs for Across. Since Event extends Log, it's safe to narrow the
scope of the RPC data down to a Log type. This significantly reduces the
delta to viem's Log type and is a key stepping stone for viem and ethers
to be used interchangeably for RPC interfacing & transport.

It additionally seems like this simplification _might_ also support
Solana events (...with some eventual coercion).