improve(SpokePoolClient): Skip unnecessary LP fee computation (#736)
The SpokePoolClient currently piggybacks on the HubPoolClient's batch LP
fee computation in order to optimise resolving quoteTimestamp to a
mainnet block number. This was historically necessary prior to the v3
transition, but it's no longer needed now that the LP fee is computed by
upper-layer code. Dropping the LP fee computation will reduce RPC 
queries on update.