experimental: gasPriceOracle overhaul
This change updates the gasPriceOracle to leverage viem's gas pricing
strategies. This is beneficial because viem has an active community that
are actively adding support for new chains and improving support for
existing chains. An example of this is Linea, where linea_estimateGas is
supported out of the box after a recent update.

In order to maintain some degree of compatibility with existing users of
the gasPriceOracle, the external interface still accepts an ethers
Provider instance. It then internally instantiates a viem PublicClient
instance and uses that instead. There is still a residual issue of how
to resolve the relevant provider URLs, since it's not necessarily
reliable to pull from the provider due to the uncertain type (i.e.
StaticJsonRpcProvider or one of the extended variants that wraps
multiple providers).