improve(SpokePoolClient): Multicall latestBlock updates
The SpokePoolClient is currently throwing semi-regular errors on update.
This seems to be particularly occurring against Polygon. The underlying
reason for this is that the RPC quroum system is performing an eth_call
against multiple providers, tagged at the latest block number as
resolved by only one of them. Some providers may not have resolved that
block number yet, in which case the query fails.

The instance of this can presumably be reduced by bundling the two calls
where this occurs into a single SpokePool multicall. As a side-benefit,
it will reduce RPC consumption.