refactor: Remove SpokePoolClient oldestTime lookup
The `oldestTime` attribute is only needed for a very specific check when
validating block ranges in the dataworker. Removing it from the standard
SpokePoolClient update flow is a simplification for non-dataworker
use-cases - i.e. relayer, monitor, finalizer. The dataworker can instead
resolve the timestamp of the searchConfig fromBlock on demand.

The underlying motivation for this change is to simplify the interface
between the relayer and the external SpokePool listener processes.