fix(RetryProvider): Avoid unnecessary promise reassignment
Debug logging has shown that the existing logic unconditionally
completes the for() loop `this.retries` times for each RPC query. This
doesn't seem to submit `this.retries` queries but does result in a lot
of unnecessary promise reassignment, which might be wasting CPU and RAM
in a tight loop.

The structure of the change here is made in anticipation of a follow-up
commit to introduce some intelligence into the retry loop to avoid
retries when the failure was actually "legitimate" (i.e. eth_call
revert).