fix(RetryProvider): Avoid unnecessary promise reassignment (#761)
Debug logging has shown that the existing logic unconditionally
completes the for() loop this.retries times for each RPC query. This
doesn't seem to submit this.retries queries but does result in a lot of
unnecessary promise reassignment.

The structure of the change here is made in anticipation of a follow-up
commit to introduce some intelligence into the retry loop to avoid
retries when the failure was actually "legitimate" (i.e. eth_call
revert). Without this change it was proving very difficult to bail out 
of the retry loop due to the promise chaining that implies the use of
closures.