improve(PriceClient): Add default price adapter
This adapter can be instantiated by the caller to permit the PriceClient
to default to a price of 0 for any token that it was unable to resolve
via the preceding adapters. This therefore provides the caller with a
simple way of controlling whether a token price resolution failure
should throw or not.