improve(PriceClient): Add default price adapter (#774)
This adapter can be instantiated by the caller to permit the PriceClient
to default to a price of 0 for any token that it was unable to resolve
via the preceding adapters. It can additionally be automatically
instantiated by the PriceClient based on an input option being set. This
therefore provides the caller with a simple way of controlling whether a
token price resolution failure should throw or not. It additionally
provides a pathway for the caller to be able to specify fixed prices for 
one or more tokens.