refactor(SpokePoolClient): Move findDeposit to helper function file
This function `findDeposit` is only used by `queryHistoricalDepositForFill`, a utility function. I propose moving this helper function closer to the helper function as it doesn't need to be a client level function. Though, it does make lots of spoke pool client calls but I see it just as intuitive to pass in the relevant spoke pool client into the function. The function effecitvely "finds a deposit" using a spoke pool client's memory.