improve(gasPriceOracle): Dynamic priority fee on EIP 1559 queries (#803)
This change introduces a new "raw" variant of the gasPriceOracle
EIP-1559 adadpter. This new adapter calls the eth_maxPriorityFeePerGas
method in parallel with querying the gas price of the pending block.
This is intended as a replacement for the ethers getFeeData() query,
which resolves the base fee by querying the latest block, but annoyingly
hardcodes the priority fee to 1.5 Gwei.

Testing on Optimism shows that this returns a fixed fee of 0xf4240. This
is however returned by the RPC endpoint, rather than determined by
ethers itself.

Testing on ethereum also shows a lower priority fee vs. vanilla ethers
v5 with its fixed 1.5 Gwei priority fee.

The raw adapter can be enabled via setting the environment variable:
GAS_PRICE_EIP1559_RAW_=true