refactor(utils): Redirect all gas price queries to gasPriceOracle (#805)
Rearrange this utility function to force gas price queries via the 
gasPriceOracle. It has been verified that it's OK to bypass the OP sdk 
in this case, since it only queries provider.gasPrice() anyway.

nb. This is an interim refactor and is not the end result. Ultimately
the gasPriceOracle will need to accept a transaction object and will be
able to give an estimate back. This is necessary for Linea, but can also
be (ab)used to simplify the OP stack workflow for estimating gas 
consumption.