feat(GasPriceOracle): Allow caller to set a priority fee multiplier
This acts similarly to the baseFeeMultiplier. It is unused for `legacy()` pricing which returns a single gas price so we use the `baseFeeMultiplier` to scale the gas price. It's also unused for the Linea and Arbitrum adapters which only have one component of the gas price that is variable (the priority fee and base fee, respectively) and we use the `baseFeeMultiplier` in both cases.